pyproject.toml
[project]
name = "scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.8.1",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "django==5.2.5",
    "dnspython==2.7.0",
    "email-validator==2.3.0",
    "fastapi[standard]==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx[http2]==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "langchain-groq>=0.3.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mcp-use>=1.3.7",
    "mdurl==0.1.2",
    "numpy==2.3.2",
    "pip==25.2",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "rich==14.1.0",
    "rich-toolkit==0.15.0",
    "rignore==0.6.4",
    "sentry-sdk==2.35.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sqlmodel==0.0.24",
    "sqlparse==0.5.3",
    "starlette==0.47.3",
    "typer==0.16.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "sentence-transformers>=5.2.0",
    "qdrant-client>=1.16.2",
    "transformers>=4.57.3",
    "torch>=2.9.1",
    "scikit-learn>=1.8.0",
    "pandas>=2.3.3",
    "joblib>=1.5.3",
    "sympy>=1.14.0",
    "mpmath>=1.3.0",
    "huggingface-hub>=0.36.0",
    "grpcio>=1.76.0",
    "protobuf>=6.33.2",
]
